Pundit

+++ app/controllers/orders_controller.rb
@@ -1,11 +1,13 @@
 class OrdersController < ApplicationController
   def new
-    @unconfirmed_orders = Order.where(status: false)
+    @unconfirmed_orders = Order.where(status: false, user: current_user)
+    authorize @unconfirmed_orders
   end
 
   def create
     order = Order.new(customer: Customer.first)
     order.user = current_user
+    authorize order
     order.status = false
     if order.save!
       redirect_to order_list_orders_path(order)
@@ -14,6 +16,7 @@ class OrdersController < ApplicationController
 
   def confirm
     order = Order.find(params[:id])
+    authorize order
     decrease_stock(order)
     order.status = true
     order.save
@@ -25,6 +28,7 @@ class OrdersController < ApplicationController
 
   def destroy
     order = Order.find(params[:id])
+    authorize order
     order.destroy
     redirect_to new_order_path
   end
