+++ app/controllers/purchases_controller.rb
@@ -2,13 +2,15 @@ class PurchasesController < ApplicationController
 
   def new
     @purchase = Purchase.new
-    @unconfirmed_purchases = Purchase.where(status: false)
+    @unconfirmed_purchases = Purchase.where(status: false, user: current_user)
     @supplier = Supplier.new
+    authorize @purchase
   end
 
   def create
     purchase = Purchase.new(purchase_params)
     purchase.user = current_user
+    authorize purchase
     purchase.status = false
     if purchase.save
       redirect_to purchase_list_purchases_path(purchase)
@@ -19,16 +21,19 @@ class PurchasesController < ApplicationController
     purchase = Purchase.find(params[:id])
     increase_stock(purchase)
     purchase.status = true
+    authorize purchase
     purchase.save
     redirect_to purchase_path(purchase)
   end
 
   def show
     @purchase = Purchase.find(params[:id])
+    authorize @purchase
   end
 
   def destroy
     purchase = Purchase.find(params[:id])
+    authorize purchase
     purchase.destroy
     redirect_to new_purchase_path
   end
